back-end worked on

+++ teams/views.py
@@ -1,5 +1,4 @@
 # pylint: disable=no-member
-from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.status import HTTP_404_NOT_FOUND, HTTP_201_CREATED, HTTP_422_UNPROCESSABLE_ENTITY, HTTP_204_NO_CONTENT, HTTP_202_ACCEPTED, HTTP_401_UNAUTHORIZED
@@ -20,6 +19,7 @@ class TeamListView(APIView):
 
     def post(self, request):
 
+        request.data['manager'] = request.user.id
         team = TeamSerializer(data=request.data)
         if team.is_valid():
             team.save()
@@ -42,8 +42,11 @@ class TeamDetailView(APIView):
 
     def put(self, request, pk):
 
+        request.data['manager'] = request.user.id
         try:
             team = Team.objects.get(pk=pk)
+            if team.manager.id != request.user.id:
+                return Response(status=HTTP_401_UNAUTHORIZED)
             updated_team = TeamSerializer(team, data=request.data)
             if updated_team.is_valid():
                 updated_team.save()
@@ -52,10 +55,12 @@ class TeamDetailView(APIView):
         except:
             return Response({'message': 'Not Found'}, status=HTTP_404_NOT_FOUND)
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
 
         try:
             team = Team.objects.get(pk=pk)
+            if team.manager.id != request.user.id:
+                return Response(status=HTTP_401_UNAUTHORIZED)
             team.delete()
             return Response(status=HTTP_204_NO_CONTENT)
         except Team.DoesNotExist:
